:sparkles: Integrate HasContainerContentSize
