+++ Sources/SwiftTUI/View/Stack/VStack.swift
@@ -46,38 +46,4 @@ extension VStack: ViewGraphSetAcceptable {
 }
 
 extension _VStackLayout: VariadicView.Root { }
-
-extension VStack: HasContainerContentSize {
-    func containerContentSize(viewGraph: ViewGraph, visitor: ViewSetRectVisitor) -> Size {
-        let rendableChildren: [ViewGraph]
-        switch viewGraph.children.first {
-        case .some(let view) where view.anyView is _TupleView:
-            rendableChildren = view.rendableChildren
-        case _:
-            rendableChildren = viewGraph.rendableChildren
-        }
-        
-        var allocableHeight: PhysicalDistance = viewGraph.proposedSize.height - (rendableChildren.count - 1) * viewGraph.spacing
-        var maxElementWidth: PhysicalDistance = 0
-        rendableChildren.enumerated().forEach { (offset, element) in
-            let provisionalElementHeight: PhysicalDistance = allocableHeight / (rendableChildren.count - offset)
-            let elementProposedSize = Size(width: viewGraph.proposedSize.width, height: max(provisionalElementHeight, 0))
-            element.proposedSize = elementProposedSize
-            element.accept(visitor: visitor)
-            
-            maxElementWidth = max(maxElementWidth, element.rect.size.width + element.rect.origin.x)
-            allocableHeight -= element.rect.size.height
-        }
-        
-        maxElementWidth = min(maxElementWidth, viewGraph.proposedSize.width)
-        
-        switch allocableHeight {
-        case let allocableHeight where allocableHeight < 0:
-            return Size(width: maxElementWidth, height: viewGraph.proposedSize.height + abs(allocableHeight))
-        case let allocableHeight where allocableHeight > 0:
-            return Size(width: maxElementWidth, height: viewGraph.proposedSize.height - allocableHeight)
-        case _:
-            return Size(width: maxElementWidth, height: viewGraph.proposedSize.height)
-        }
-    }
-}
+extension VStack: HasContainerContentSize { }
